Menu Created
Started very late in the night. Is barely functional, and is using an asset from the unity store, which is probably my worst investment of all year, that doesn't do much.
Needs work on settings, but at least works as a main menu.
It's 6:15am, am running on two hours of sleep, [redacted] thoughts and an Oreo, and I'm ready to fight god, OR BECOME HIM